Extract blog types into apps/blog/types.ts.
Keep utils focused on post loading helpers and update imports across blog components.
