+++ apps/blog/components/post-header/post-header.tsx
@@ -1,6 +1,6 @@
 import { User } from "lucide-react"
 import { PostCover } from "@/blog/components/post-cover"
-import type { BlogPostAuthor } from "@/blog/utils"
+import type { BlogPostAuthor } from "@/blog/types"
 import { Link } from "@/ui/link"
 import { Heading, Small } from "@/ui/typography"
