+++ apps/blog/post-modules.ts
@@ -1,4 +1,4 @@
-import type { BlogPostMdxModule } from "./utils"
+import type { BlogPostMdxModule } from "./types"
 
 export const postModules = import.meta.glob<BlogPostMdxModule>(
   "./content/*.mdx",
